+++ x-uploader/lib/x/uploader.rb
@@ -1,5 +1,6 @@
 require_relative "uploader/version"
 require_relative "uploader/account"
+require_relative "uploader/api"
 require_relative "uploader/gif"
 require_relative "uploader/invalid_media_type"
 require_relative "uploader/media"